# api/openapi.yaml
openapi: 3.1.1
info:
  title: Bulktrack API
  version: 1.0.0
  description: |
    API definition for the BulkTrack application.
servers:
  - url: https://api.bulk-track.com/v1
    description: Production server
  - url: http://localhost:8787/v1
    description: Local development server

security:
  - BearerAuth: []

paths:
  # ---------------- Auth ---------------- #
  /v1/auth/device:
    post:
      summary: Activate a new device (anonymous onboarding)
      tags: [Auth]
      parameters:
        - name: X-Device-Id
          in: header
          required: true
          description: UUID v7 of the device
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Device tokens issued, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
      security: []

  /v1/auth/apple:
    post:
      summary: Link or sign-in with Apple
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identityToken]
              properties:
                identityToken: { type: string, description: Apple identity token (JWT) }
      responses:
        '200': { description: User tokens, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
      security: []

  /v1/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '200': { description: Tokens refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/TokenResponse' } } } }
      security: []

  /v1/auth/logout:
    post:
      summary: Log out (revoke refresh token)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshTokenRequest' }
      responses:
        '204': { description: Logged out }
      security: []

  # ---------------- Exercises ---------------- #
  /v1/exercises:
    get:
      summary: Search available exercises or list all available exercises.
      tags: [Exercises]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - in: query
          name: q
          schema: { type: string }
          description: Prefix search string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Maximum number of exercises to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Offset for pagination.
      responses:
        '200': { description: List, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Exercise' } } } } }
    post:
      summary: Create custom exercise
      tags: [Exercises]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExerciseCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Exercise' } } } }

  /v1/me/exercises/recent:
    get:
      summary: List recently used exercises for the authenticated user
      tags: [Exercises]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Maximum number of exercises to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Offset for pagination.
      responses:
        '200':
          description: A list of recently used exercises.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exercise'
        '401':
          description: Unauthorized
      security:
        - BearerAuth: []

  # ---------------- Sets ---------------- #
  /v1/sets:
    post:
      summary: Add set
      tags: [Sets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetCreate' }
      responses:
        '201': { description: Recorded, content: { application/json: { schema: { $ref: '#/components/schemas/WorkoutSet' } } } }
        '400': { description: Invalid request body or parameters, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      security:
        - BearerAuth: []
    get:
      summary: List sets for the authenticated user
      tags: [Sets]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: date
          schema: { type: string, format: date }
          description: Filter sets for a specific date (YYYY-MM-DD).
        - in: query
          name: exerciseId
          schema: { type: string, format: uuid }
          description: Filter sets by exercise ID.
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [performedAt_asc, performedAt_desc]
            default: performedAt_desc
          description: "Sort order for the sets. Defaults to 'performedAt_desc' (most recent first)."
      responses:
        '200': { description: List, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/WorkoutSet' } } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      security:
        - BearerAuth: []

  /v1/sets/{setId}:
    patch:
      summary: Update a specific workout set
      tags: [Sets]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - name: setId
          in: path
          required: true
          description: ID of the set to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUpdate'
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/WorkoutSet' } } } }
        '400': { description: Invalid request body or parameters, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Set not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      security:
        - BearerAuth: []
    delete:
      summary: Delete a specific workout set
      tags: [Sets]
      parameters:
        - name: setId
          in: path
          required: true
          description: ID of the set to delete
          schema:
            type: string
            format: uuid
      responses:
        '204': { description: Set deleted successfully }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Forbidden, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Set not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
      security:
        - BearerAuth: []

  # ---------------- Dashboard ---------------- #
  /v1/dashboard:
    get:
      summary: Fetch dashboard aggregates
      tags: [Dashboard]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for translatable strings like muscle group names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - in: query
          name: span
          schema:
            type: string
            enum: [1w, 4w, 8w, 12w, 24w]
            default: 1w
          description: Week span window (inclusive, ending current ISO week).
      responses:
        '200':
          description: Dashboard payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardResponse' }

  # ---------------- Workout History ---------------- #
  /v1/me/workouts:
    get:
      summary: List daily workout summaries for the authenticated user
      tags: [Workouts]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Maximum number of days to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Offset for pagination.
      responses:
        '200':
          description: List of daily workout summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailyWorkoutSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []

  /v1/me/workouts/{date}:
    get:
      summary: Get detailed workout data for a specific date
      tags: [Workouts]
      parameters:
        - name: Accept-Language
          in: header
          required: false
          description: "Preferred language for exercise names and muscle group names (e.g., en, ja). If not specified, the server's default language will be used."
          schema: { type: string }
        - name: date
          in: path
          required: true
          description: Date to retrieve workout details for (YYYY-MM-DD format)
          schema:
            type: string
            format: date
            examples: ['2025-05-29']
      responses:
        '200':
          description: Detailed workout data for the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyWorkoutDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No workout data found for the specified date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Global Bearer auth

  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: ID of the session
      schema:
        type: string
        format: uuid

  # ---- Schemas (same as previous version, trimmed) ---- #
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          description: An error code identifying the type of error.
        message:
          type: string
          description: A human-readable message providing more details about the error.
      required:
        - code
        - message
    TokenResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
      required: [accessToken, refreshToken, expiresIn]
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken: { type: string }
      required: [refreshToken]
    UserCreate:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string }
    User:
      allOf:
        - $ref: '#/components/schemas/UserCreate'
        - type: object
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
    DeviceCreate:
      type: object
      required: [userId]
      properties:
        userId: { type: string, format: uuid }
        platform: { type: string }
    Device:
      allOf:
        - $ref: '#/components/schemas/DeviceCreate'
        - type: object
          properties:
            deviceId: { type: string, format: uuid }
            linkedAt: { type: string, format: date-time }
    ExerciseCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
    Exercise:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        isOfficial: { type: boolean, description: "Indicates if the exercise is an official (system-defined) one or a user-custom one. True for official, false for custom." }
        lastUsedAt: { type: [string, "null"], format: date-time }
        useCount: { type: [integer, "null"] }
    SessionStart:
      type: object
      required: [startedAt]
      properties:
        startedAt: { type: string, format: date-time }
        notes: { type: [string, "null"] }
    WorkoutSession:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, "null"], format: date-time }
        notes: { type: [string, "null"] }
    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/WorkoutSession'
        - type: object
          properties:
            sets:
              type: array
              items: { $ref: '#/components/schemas/WorkoutSet' }
    SetCreate:
      type: object
      required:
        - exerciseId
        - weight
        - reps
        - performedAt
      properties:
        exerciseId:
          type: string
          format: uuid
          description: ID of the exercise for this set.
        weight:
          type: number
          format: float
          description: Weight used for the set (e.g., in kg or lbs).
        reps:
          type: integer
          description: Number of repetitions performed.
        rpe:
          type: [number, "null"]
          format: float
          minimum: 1
          maximum: 10
          description: Rating of Perceived Exertion (1-10).
        notes:
          type: [string, "null"]
          description: Additional notes for the set.
        performedAt:
          type: string
          format: date-time
          description: Timestamp when the set was performed.
    WorkoutSet:
      type: object
      properties:
        id: { type: string, format: uuid }
        exerciseId: { type: string, format: uuid }
        setNumber: { type: integer, description: "The order of the set in the session." }
        weight: { type: number, format: float }
        reps: { type: integer }
        rpe: { type: [number, "null"], format: float, minimum: 1, maximum: 10 }
        notes: { type: [string, "null"] }
        performedAt: { type: string, format: date-time }
    SetUpdate:
      type: object
      properties:
        exerciseId:
          type: [string, "null"]
          format: uuid
          description: ID of the exercise for this set.
        weight:
          type: [number, "null"]
          format: float
          description: Weight used for the set (e.g., in kg or lbs).
        reps:
          type: [integer, "null"]
          description: Number of repetitions performed.
        rpe:
          type: [number, "null"]
          format: float
          minimum: 1
          maximum: 10
          description: Rating of Perceived Exertion (1-10).
        notes:
          type: [string, "null"]
          description: Additional notes for the set.
        performedAt:
          type: [string, "null"]
          format: date-time
          description: Timestamp when the set was performed.
    WeekPoint:
      type: object
      required: [weekStart, totalVolume]
      properties:
        weekStart: { type: string, format: date, examples: ['2025-05-05'] }
        totalVolume: { type: number, examples: [12500] }
        avgSetVolume: { type: number, examples: [180] }
        e1rmAvg: { type: [number, "null"], examples: [110] }

    MuscleGroupSeries:
      type: object
      required: [muscleGroupId, groupName, points]
      properties:
        muscleGroupId: { type: integer }
        groupName: { type: string }
        points:
          type: array
          items: { $ref: '#/components/schemas/MuscleGroupWeekPoint' }

    MuscleGroupWeekPoint:
      type: object
      required: [weekStart, totalVolume, setCount]
      properties:
        weekStart: { type: string, format: date, examples: ['2025-05-05'] }
        totalVolume: { type: number, examples: [2500] }
        setCount: { type: integer, examples: [20] }
        avgE1rm: { type: [number, "null"], examples: [80.0] }

    MetricSeries:
      type: object
      required: [metricKey, unit, points]
      properties:
        metricKey: { type: string, examples: ["body_weight"] }
        unit: { type: string, examples: ["kg"] }
        points:
          type: array
          items:
            type: object
            required: [weekStart, value]
            properties:
              weekStart: { type: string, format: date, examples: ['2025-05-05'] }
              value: { type: number, examples: [70.4] }

    DashboardResponse:
      type: object
      required: [thisWeek, lastWeek, trend, muscleGroups, metrics]
      properties:
        thisWeek: { $ref: '#/components/schemas/WeekPoint' }
        lastWeek: { $ref: '#/components/schemas/WeekPoint' }
        trend:
          type: array
          items: { $ref: '#/components/schemas/WeekPoint' }
          description: Bar/line data for the requested span.
        muscleGroups:
          type: array
          items: { $ref: '#/components/schemas/MuscleGroupSeries' }
        metrics:
          type: array
          items: { $ref: '#/components/schemas/MetricSeries' }

    DailyWorkoutSummary:
      type: object
      required: [date, exercises, totalVolume, avgRM, setCount]
      properties:
        date:
          type: string
          format: date
          description: The date of the workout (YYYY-MM-DD format)
          examples: ['2025-05-29']
        exercises:
          type: array
          items:
            type: object
            required: [exerciseId, exerciseName]
            properties:
              exerciseId: { type: string, format: uuid }
              exerciseName: { type: string, description: "Exercise name in the requested language" }
          description: List of exercises performed on this date
        totalVolume:
          type: number
          format: float
          description: Total training volume for the day (weight × reps)
          examples: [12500]
        avgRM:
          type: number
          format: float
          description: Average 1RM estimation for the day
          examples: [85.5]
        setCount:
          type: integer
          description: Total number of sets performed on this date
          examples: [25]

    DailyWorkoutDetail:
      type: object
      required: [date, exercises, totalVolume, avgRM]
      properties:
        date:
          type: string
          format: date
          description: The date of the workout (YYYY-MM-DD format)
          examples: ['2025-05-29']
        totalVolume:
          type: number
          format: float
          description: Total training volume for the day
          examples: [12500]
        avgRM:
          type: number
          format: float
          description: Average 1RM estimation for the day
          examples: [85.5]
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseDetail'

    ExerciseDetail:
      type: object
      required: [exerciseId, exerciseName, sets, totalVolume, avgRM, muscleVolumeBreakdown]
      properties:
        exerciseId:
          type: string
          format: uuid
        exerciseName:
          type: string
          description: Exercise name in the requested language
        sets:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutSet'
          description: All sets performed for this exercise on this date
        totalVolume:
          type: number
          format: float
          description: Total volume for this exercise (weight × reps)
          examples: [3500]
        avgRM:
          type: number
          format: float
          description: Average 1RM estimation for this exercise
          examples: [90.0]
        muscleVolumeBreakdown:
          type: array
          items:
            type: object
            required: [muscleGroupId, muscleGroupName, effectiveVolume]
            properties:
              muscleGroupId:
                type: integer
                description: ID of the muscle group
              muscleGroupName:
                type: string
                description: Name of the muscle group in the requested language
              effectiveVolume:
                type: number
                format: float
                description: Effective volume for this muscle group (volume × tension_ratio × tension_factor)
                examples: [2100]